D4PG_streaming_code/data_streaming/EMG/Cyguns python src/simple-viewer/requirements.txt
matplotlib==3.5.3
numpy==1.23.3
websocket_client==1.4.1
